% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysis}
\alias{analysis}
\title{Run General Analysis on Several Experiments}
\usage{
analysis(
  data_patient_annotated = c(),
  data_control_annotated = c(),
  genotype_matrix = c(),
  KindOfRegion = c(),
  filtering_criteria = list(),
  filtering_level = list(),
  filtering_direction = list(),
  min_var = 2,
  max_var = 10000,
  remove_private = FALSE,
  over_represented_pathway = c(),
  version = "optimal",
  var_covariate = 0,
  var_covariate_pca = 1,
  rare_maf_threshold = 0.01,
  pvalue_threshold = 0.05,
  Binary = TRUE,
  path_results = "",
  path_general_results = ""
)
}
\arguments{
\item{data_patient_annotated}{A data frame containing annotated patient data.}

\item{data_control_annotated}{A data frame containing annotated control data.}

\item{genotype_matrix}{A matrix representing the genotype data.}

\item{KindOfRegion}{A vector indicating the regions of interest for the analysis.}

\item{filtering_criteria}{A list specifying filtering criteria for the data.}

\item{filtering_level}{A list specifying the level of filtering to apply.}

\item{filtering_direction}{A list indicating the filtering direction.}

\item{min_var}{Integer. The minimum number of variants required for a region to be analyzed. Default is 2.}

\item{max_var}{Integer. The maximum number of variants allowed for a region to be analyzed. Default is 10000.}

\item{remove_private}{Logical. If TRUE, removes private variants from the analysis. Default is FALSE.}

\item{over_represented_pathway}{A list or vector indicating pathways that are over-represented.}

\item{version}{Character. Specifies the version of the statistical framework to use. Default is "optimal".}

\item{var_covariate}{Numeric. A variable used for covariate adjustment.}

\item{var_covariate_pca}{Numeric. A variable used for PCA-based covariate adjustment.}

\item{rare_maf_threshold}{Numeric. The minor allele frequency threshold for defining rare variants. Default is 0.01.}

\item{pvalue_threshold}{Numeric. The p-value threshold for statistical significance. Default is 0.05.}

\item{Binary}{Logical. If TRUE, applies a binary framework for analysis. Default is TRUE.}

\item{path_results}{Character. Path to save detailed results of each experiment.}

\item{path_general_results}{Character. Path to save the general results summary.}
}
\value{
A list containing the results of the analyses, including general experiment information and plots.
}
\description{
This function performs various genetic analyses across multiple experiments, incorporating filtering,
hypothesis generation, and statistical testing. It allows for customization of input data, filtering criteria, analysis types, and output storage.
}
\examples{
# Example usage:
}
