% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFormat.R
\name{checkFormat}
\alias{checkFormat}
\title{Check Format of Data}
\arguments{
\item{patient}{A data frame containing at least the following columns:
\itemize{
\item \code{patient}: character
\item \code{chromosome}: character
\item \code{position}: integer
\item \code{reference}: character
\item \code{alternative}: character
\item \code{zygosity}: integer
\item \code{gene_symbol}: character
}}

\item{control}{A data frame with the same required columns as \code{patient}.}
}
\value{
A list containing the two cleaned and formatted data frames:
\itemize{
\item \code{patient}: Cleaned and formatted patient data frame.
\item \code{control}: Cleaned and formatted control data frame.
}
}
\description{
Ensures that the input data frames have the desired format.
It checks for the presence of required columns, removes rows with missing values in key columns,
and standardizes chromosome values (e.g., converting "23" and "24" to "X" and "Y").
}
