% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_option.R
\name{filtering_option}
\alias{filtering_option}
\title{Generate Filtering Options for Multiple Experiments}
\usage{
filtering_option(
  KindOfRegion = list(),
  filtering_criteria = list(),
  filtering_level = list()
)
}
\arguments{
\item{KindOfRegion}{A character vector specifying the type(s) of genetic regions to analyze. Possible values include "gene", "pathway", "GO", "over_represented_pathway", "over_represented_GO", and combinations thereof.}

\item{filtering_criteria}{A list of filtering criteria to apply for each experiment.}

\item{filtering_level}{A list of filtering levels corresponding to each filtering criterion.}
}
\value{
A list containing:
\itemize{
\item \code{vec_hypotheses}: A data frame summarizing filtering criteria, levels, and genetic regions for each experiment.
\item \code{Nbr_exp}: The total number of experiments generated based on input parameters.
\item \code{KindOfRegion}: The processed vector of genetic region types used in the analysis.
}
}
\description{
This function prepares filtering options based on criteria, levels, and types of genetic regions for multiple experiments. It supports dynamic combinations of filtering criteria and levels for different experiment types such as genes, pathways, and Gene Ontology (GO) terms.
}
\examples{
# Example 1: Simple filtering setup for genes
filtering_option(
  KindOfRegion = "gene",
  filtering_criteria = list("criterion1"),
  filtering_level = list("level1")
)

# Example 2: Multiple filtering options for genes and pathways
filtering_option(
  KindOfRegion = c("gene", "pathway"),
  filtering_criteria = list("criterion1", "criterion2"),
  filtering_level = list("level1", "level2")
)

}
