% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_covered_matrix.R
\name{genotype_covered_matrix}
\alias{genotype_covered_matrix}
\title{Generate a Cleaned Genotype and Coverage Matrix}
\arguments{
\item{coverage_matrix}{A data frame representing the coverage matrix, with positions in rows and individuals in columns.}

\item{genotype_matrix}{A data frame representing the genotype matrix, with positions in rows and individuals in columns.}

\item{variant}{A data frame containing variant information, including columns \code{id} and \code{position}.}

\item{patient_name}{A vector of patient sample names corresponding to columns in the matrices.}

\item{control_name}{A vector of control sample names corresponding to columns in the matrices.}

\item{min_coverage}{An integer specifying the minimum coverage threshold for a position to be considered reliable. Default is 20.}

\item{remove_individual}{A logical value indicating whether to remove individuals with a high number of poorly covered positions. Default is TRUE.}

\item{remove_position}{A logical value indicating whether to remove positions with a high number of poorly covered individuals. Default is TRUE.}

\item{path_store_result}{A character string specifying the directory path to store result files (e.g., plots, diagnostics).}
}
\value{
A list containing cleaned genotype and coverage matrices, filtered variants, rejected variants, and (if \code{remove_individual} is TRUE) information about rejected individuals.
}
\description{
This function processes genotype and coverage matrices to clean and filter the data based on variant coverage, removing low-quality individuals and positions. It also calculates data reliability and outputs diagnostic plots and cleaned datasets.
}
