% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeMatrix.R
\name{genotypeMatrix}
\alias{genotypeMatrix}
\title{Construct Genotype Matrix}
\arguments{
\item{data_patient}{A data frame containing the columns: sample, chr, pos, reference, alternative, zygosity.}

\item{data_control}{A data frame containing the columns: sample, chr, pos, reference, alternative, zygosity.}

\item{Nbr_patient_init}{The initial number of patients expected (used for matrix dimensioning).}

\item{Nbr_control_init}{The initial number of controls expected (used for matrix dimensioning).}
}
\value{
A list containing the genotype matrix and three lists of variant identifiers: those specific to patients,
those specific to controls, and a combined list from both sets.
}
\description{
This function constructs a genotype matrix from patient and control data. The function expects data frames
containing genetic variant information such as chromosome position, reference, and alternate alleles,
as well as zygosity. The output is a genotype matrix where columns represent variants and rows represent individuals.
Additional output includes lists of unique variants specific to patients and controls.
}
