% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvcf_pretreatment.R
\name{gvcf_pretreatment}
\alias{gvcf_pretreatment}
\title{Pre-treat VCF Data}
\arguments{
\item{path_data}{The file path to the input gVCF data.}

\item{path_output}{The output directory for saving processed files and reports.}

\item{cases}{The file path to the list of case samples.}

\item{controls}{The file path to the list of control samples.}

\item{mandatory_column_mapping}{A list mapping gVCF column names to required names for processing.}

\item{MAF}{The minor allele frequency threshold for variant filtering.}

\item{het_var}{A vector of genotype codes representing heterozygous variants.}

\item{homo_var}{A vector of genotype codes representing homozygous variants.}

\item{homo_ref}{A vector of genotype codes representing homozygous reference.}

\item{cleaning}{A boolean flag to perform data cleaning based on coverage and variant quality.}

\item{remove_individual}{A boolean flag to remove individuals based on poor coverage across many positions.}

\item{remove_position}{A boolean flag to remove poorly covered positions across many individuals.}

\item{min_coverage}{The minimum coverage required for a position to be considered reliable.}
}
\value{
A list containing two data frames: one for case data and one for control data after preprocessing.
}
\description{
This function preprocesses VCF (and gVCF) data for variant analysis. It reads the gVCF file, renames columns according
to a provided mapping, and filters variants based on the minor allele frequency (MAF). It also identifies and renames genes,
performs data cleaning based on coverage and allele frequency, and splits the data into cases and controls.
}
