% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypotheses_generation.R
\name{hypotheses_generation}
\alias{hypotheses_generation}
\title{Generate Hypotheses from Annotated Patient and Control Data}
\usage{
hypotheses_generation(
  data_patient_annotated = c(),
  data_control_annotated = c(),
  filtering_criteria = list(),
  filtering_level = list(),
  filtering_direction = list(),
  level_of_analysis = c("gene", "pathway", "GO")
)
}
\arguments{
\item{data_patient_annotated}{A data frame with annotated data for patients.}

\item{data_control_annotated}{A data frame with annotated data for controls.}

\item{filtering_criteria}{A list containing the criteria names to filter the datasets.}

\item{filtering_level}{A list containing the threshold levels for each filtering criterion.}

\item{filtering_direction}{A list indicating the direction of filtering ('+' for above or '-' for below the threshold).}

\item{level_of_analysis}{A vector indicating the levels of analysis ('gene', 'pathway', 'GO') to generate hypotheses.}
}
\value{
A list containing the filtered patient data, filtered control data, and a vector of regions of interest based on the level of analysis.
}
\description{
This function filters annotated datasets of patients and controls based on specified criteria and then generates hypotheses
at different levels of genetic analysis such as gene, pathway, or gene ontology (GO) terms. The function allows flexible
filtering with dynamic criteria and supports analysis at multiple biological levels, facilitating targeted hypothesis generation.
}
