% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Main Function for Genetic Data Analysis Pipeline}
\usage{
main(
  PATH_PATIENTS = "",
  PATH_CONTROLS = "",
  PATH_GVCF = "",
  mandatory_column_mapping = list(sample = "Indiv", chr = "CHROM", pos = "POS", reference
    = "REF", alternative = "ALT", zygosity = "", gene_symbol = "Gene.ensGene",
    gnomad_wes_af = "gnomad40_AF", gt_GT = "gt_GT", gt_DP = "gt_DP"),
  path_results = "",
  importation_method = 0,
  RLIBRARIES = .libPaths(),
  list_genes = 0,
  rare_maf_threshold = 0.01,
  var_covariate = 0,
  var_covariate_pca = 1,
  min_var = 2,
  max_var = 10000,
  do_data_preparation = 1,
  DO_over_representation = 1,
  version = "optimal",
  filtering_criteria = 0,
  filtering_level = 0,
  filtering_direction = 0,
  coverage_matrix_path = 0,
  mapqv_threshold = 20,
  min_coverage = 20,
  remove_individual = 1,
  remove_position = 1,
  remove_private = 0,
  pvalue_threshold = 0.05,
  Binary = TRUE,
  path_Excalibur_function = "/Users/sboutry/Documents/excalibur/Rpackage/DoAggregate/R"
)
}
\arguments{
\item{PATH_PATIENTS}{File path to the patient data.}

\item{PATH_CONTROLS}{File path to the control data.}

\item{PATH_GVCF}{File path to genomic VCF data.}

\item{mandatory_column_mapping}{A list mapping column names in the input data to required format.}

\item{path_results}{Directory path to store results.}

\item{importation_method}{Integer indicating the method of data importation (0 for default, 1 for GVCF).}

\item{RLIBRARIES}{Paths to R libraries used in the analysis.}

\item{list_genes}{File path to a list of genes of interest, if any.}

\item{rare_maf_threshold}{Threshold for defining rare minor allele frequency.}

\item{var_covariate}{Integer indicating whether to use covariates in the analysis.}

\item{var_covariate_pca}{Boolean indicating whether to perform PCA on covariates.}

\item{min_var}{Minimum number of variants to include for analysis.}

\item{max_var}{Maximum number of variants to consider in the analysis.}

\item{do_data_preparation}{Boolean indicating whether to perform data preparation.}

\item{DO_over_representation}{Boolean indicating whether to perform over-representation analysis.}

\item{version}{String indicating the version of the analysis method.}

\item{filtering_criteria}{List of criteria for filtering the data.}

\item{filtering_level}{List of levels corresponding to each filtering criterion.}

\item{filtering_direction}{List indicating the direction of each filtering criterion.}

\item{coverage_matrix_path}{Optional path to a pre-existing coverage matrix.}

\item{mapqv_threshold}{Mapping quality threshold.}

\item{min_coverage}{Minimum coverage threshold.}

\item{remove_individual}{Boolean indicating whether to remove individuals based on poor data quality.}

\item{remove_position}{Boolean indicating whether to remove positions based on coverage.}

\item{remove_private}{Boolean indicating whether to exclude private variants.}

\item{pvalue_threshold}{P-value threshold for statistical tests.}

\item{Binary}{Boolean indicating whether the analysis is binary.}

\item{path_Excalibur_function}{Path to the Excalibur function library.}
}
\value{
No return value, side effects include writing results to files and possibly printing to console.
}
\description{
This comprehensive function orchestrates the entire workflow for genetic data analysis. It imports data,
applies pre-treatment steps, generates genotype matrices, conducts data preparation, and carries out
over-representation analysis. It is designed to handle both patient and control data, apply filtering based
on specified criteria, and execute advanced genetic analyses including PCA and hypothesis testing.
}
