% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/over_representation_analysis.R
\name{over_representation_analysis}
\alias{over_representation_analysis}
\title{Over Representation Analysis for Gene Lists}
\arguments{
\item{geneList}{Vector of gene symbols to analyze.}

\item{gene_annot}{Annotation type of the genes provided (default: "SYMBOL").}

\item{background_gene}{Background gene set for normalization in the analysis.}

\item{go_of_interest}{Vector specifying which gene ontology categories to consider ("BP", "MF", "CC").}

\item{max_GO_similarity_accepted}{Maximum similarity threshold for considering GO terms as distinct.}

\item{cutoff}{Significance cutoff for enrichment results.}

\item{max_gene_per_item}{Maximum number of genes allowed per GO/pathway term before filtering.}

\item{max_item_plot}{Maximum number of terms to display in plots.}

\item{coocurrence}{Whether to perform and plot gene co-occurrence analysis.}

\item{max_coocurrence}{Maximum number of top terms to include in co-occurrence analysis.}

\item{MSigDB}{Boolean to decide whether to use the MSigDB database for analysis.}

\item{path_to_table}{Path to the database tables needed for the analysis.}

\item{path_to_store_results}{Directory path where the results and plots will be saved.}
}
\value{
A list of data frames containing the over-represented pathways and GO terms along with their details.
}
\description{
This function performs over-representation analysis on provided gene lists using various databases
like MSigDB, GO, and others. It supports extensive customization for the analysis, including
background gene set adjustment, similarity cutoffs, and the ability to handle co-occurrence matrices.
The function is designed to identify significant pathways and gene ontology terms that are
over-represented in the provided gene list relative to a background set.
}
