% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_combination.R
\name{result_combination}
\alias{result_combination}
\title{Perform Comprehensive Analysis Across Experiments}
\arguments{
\item{General_info}{A data frame containing general information about each genetic region or experiment.}

\item{TrueRegion}{A vector of region names that are specifically analyzed.}

\item{summary_pvalue}{A data frame summarizing p-values for each test and region.}

\item{summary_pvalue_excell}{A data frame similar to summary_pvalue but formatted for Excel output.}

\item{Nbr_test}{The number of tests performed.}

\item{version}{Character string denoting the version of analysis, default is "optimal".}

\item{Binary}{Logical indicating whether the analysis is binary.}

\item{pvalue_threshold}{The threshold for determining significance.}

\item{level}{The level at which analysis is performed (e.g., 'gene', 'pathway').}

\item{filter_annotate_data_patient}{Filtered patient data used for annotations.}

\item{filter_annotate_data_control}{Filtered control data used for annotations.}

\item{genotype_matrix}{A matrix or data frame of genotype information.}

\item{path_store_results}{The path where results should be stored. If not provided, results are not saved.}
}
\value{
Depending on the function's processing, it may return the number of significant names detected or simply the performance metric (time taken).
}
\description{
This function performs a variety of statistical tests across multiple genetic regions or experiments,
aggregates the results, and optionally saves significant findings and summary plots.
It applies multiple testing correction and returns various statistics such as p-values and q-values.
}
