% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_region_analysis.R
\name{single_region_analysis}
\alias{single_region_analysis}
\title{Run Single Case/Control Analysis on a Specified Region}
\usage{
single_region_analysis(
  focus_data_patient = c(),
  focus_data_control = c(),
  Nbr_patient = 0,
  Nbr_control = 0,
  Nbr_test = 0,
  Binary = TRUE,
  focus_genotype_matrix = c(),
  max_var = max_var,
  var_covariate = 0,
  var_covariate_pca = FALSE,
  rare_maf_threshold = 0.01,
  pvalue_threshold = 0.05,
  version = "optimal",
  region_id = 0,
  region_of_interest = "",
  time_path = ""
)
}
\arguments{
\item{focus_data_patient}{Data frame of patient data including at least patient, chromosome, position, reference, alternative, zygosity, and gene.}

\item{focus_data_control}{Data frame of control data including similar columns as patient data.}

\item{Nbr_patient}{Number of patients.}

\item{Nbr_control}{Number of controls.}

\item{Nbr_test}{Number of tests to be performed.}

\item{Binary}{Logical indicating if the analysis is binary.}

\item{focus_genotype_matrix}{Genotype matrix focused on the specific region.}

\item{max_var}{Maximum number of variants to be considered in the PCA.}

\item{var_covariate}{Variance to be used as covariate.}

\item{var_covariate_pca}{Logical indicating whether PCA should be used for covariate adjustment.}

\item{rare_maf_threshold}{Threshold for rare minor allele frequency (MAF).}

\item{pvalue_threshold}{Threshold for significance in p-values.}

\item{version}{Specifies the version of the analysis to be used, default is 'optimal'.}

\item{region_id}{Identifier for the region being analyzed.}

\item{region_of_interest}{Name or identifier of the region of interest.}

\item{time_path}{Path to save timing data.}
}
\value{
A list containing p-values for the tests, the region of interest, and computational time.
}
\description{
This function conducts a statistical analysis on a specific genetic region using case and control data.
It processes the data based on the provided genetic matrix and performs statistical tests to determine significant variants.
}
