% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_Framework.R
\name{stat_Framework}
\alias{stat_Framework}
\title{Statistical Framework for Genetic Association Testing}
\usage{
stat_Framework(
  genotype_matrix = c(),
  phenotype = c(),
  weight_variant = c(),
  Binary = TRUE,
  covariate = 1,
  rare_maf_threshold = 0.01,
  filter.pval = 0.05,
  position = c(),
  get_test = FALSE,
  version = "optimal",
  max_var_per_test = 300
)
}
\arguments{
\item{genotype_matrix}{A matrix of genotype data where rows represent samples and columns represent genetic variants.}

\item{phenotype}{A vector of phenotypes corresponding to each sample in the genotype matrix.}

\item{weight_variant}{A vector of weights for the variants, used in weighted statistical tests.}

\item{Binary}{Logical; if TRUE, performs tests appropriate for binary traits, otherwise for continuous traits.}

\item{covariate}{Covariates to be included in the model, if any.}

\item{rare_maf_threshold}{The threshold for defining rare variants based on minor allele frequency.}

\item{filter.pval}{The p-value threshold for filtering results.}

\item{position}{A vector indicating the position of each variant, used in some types of tests.}

\item{get_test}{Logical; if TRUE, returns the names of available tests instead of performing them.}

\item{version}{Specifies the version of the tests to run ('all', 'optimal', or a specific subset).}

\item{max_var_per_test}{Maximum number of variants allowed per test to ensure computational feasibility.}
}
\value{
A list containing three elements: p-values from the tests, computational times for each test, and any errors encountered during testing.
}
\description{
This function provides a comprehensive statistical analysis for genetic association studies.
It can perform a variety of statistical tests based on the input genotype and phenotype data,
including burden tests, variance-component tests, and omnibus tests. The function allows
specifying whether to perform a binary or continuous analysis and supports adjustments for
covariates and minor allele frequency (MAF) thresholds.
}
